# Opt sum dim for middle dim case