--- a/src/flag_gems/ops/sum.py
+++ b/src/flag_gems/ops/sum.py
@@ -1,5 +1,6 @@
 import logging
 import math
+from functools import reduce
 
 import torch
 import triton
@@ -57,45 +58,6 @@ def sum_kernel_2(mid, out, mid_size, BLOCK_MID: tl.constexpr):
     tl.store(out, sum_val)
 
 
-@libentry()
-@libtuner(
-    configs=runtime.get_tuned_config("naive_reduction"),
-    key=["M", "N"],
-)
-@triton.jit
-def sum_kernel(
-    inp,
-    out,
-    M,
-    N,
-    BLOCK_M: tl.constexpr,
-    BLOCK_N: tl.constexpr,
-):
-    if tl.constexpr(inp.dtype.element_ty == tl.float16) or tl.constexpr(
-        inp.dtype.element_ty == tl.bfloat16
-    ):
-        cdtype = tl.float32
-    else:
-        cdtype = inp.dtype.element_ty
-
-    # Map the program id to the row of inp it should compute.
-    pid = tle.program_id(0) * BLOCK_M + tl.arange(0, BLOCK_M)[:, None]
-    inp = inp + pid * N
-    out = out + pid
-    row_mask = pid < M
-
-    _sum = tl.zeros([BLOCK_M, BLOCK_N], dtype=cdtype)
-    for off in range(0, N, BLOCK_N):
-        cols = off + tl.arange(0, BLOCK_N)[None, :]
-        col_mask = cols < N
-        mask = row_mask and col_mask
-
-        a = tl.load(inp + cols, mask, other=0).to(cdtype)
-        _sum += a
-    sum = tl.sum(_sum, axis=1)[:, None]
-    tl.store(out, sum, row_mask)
-
-
 def sum(inp, *, dtype=None):
     logger.debug("GEMS SUM")
     M = inp.numel()
@@ -136,8 +98,145 @@ def sum_out(inp, *, dtype=None, out):
     return out
 
 
-def sum_dim(inp, dim=None, keepdim=False, *, dtype=None):
-    logger.debug("GEMS SUM DIM")
+@libentry()
+@triton.heuristics(runtime.get_heuristic_config("softmax_non_inner"))
+@triton.jit
+def sum_dim_kernel_non_inner(
+    output_ptr,
+    input_ptr,
+    M,
+    N,
+    K,
+    TILE_N: tl.constexpr,
+    TILE_K: tl.constexpr,
+    ONE_TILE_PER_CTA: tl.constexpr,
+):
+    if tl.constexpr(input_ptr.dtype.element_ty == tl.float16) or tl.constexpr(
+        input_ptr.dtype.element_ty == tl.bfloat16
+    ):
+        cdtype = tl.float32
+    else:
+        cdtype = input_ptr.dtype.element_ty
+
+    pid_m = tle.program_id(0)
+    pid_k = tle.program_id(1)
+
+    k_offsets = pid_k * TILE_K + tl.arange(0, TILE_K)[None, :]
+
+    if ONE_TILE_PER_CTA:
+        n_offsets = tl.arange(0, TILE_N)[:, None]
+        inp_offset = pid_m * N * K + n_offsets * K + k_offsets
+        mask = (n_offsets < N) & (k_offsets < K)
+        input_ptrs = input_ptr + inp_offset
+        inp = tl.load(input_ptrs, mask=mask, other=0)
+        out = tl.sum(inp, axis=0, keep_dims=True)
+        out_offset = pid_m * K + k_offsets
+        output_ptrs = output_ptr + out_offset
+        tl.store(output_ptrs, out, mask=k_offsets < K)
+    else:
+        sum = tl.zeros([TILE_N, TILE_K], dtype=cdtype)
+
+        # specialization does not improve performance inn this example, as tested
+        for start_n in range(0, N, TILE_N):
+            n_offsets = start_n + tl.arange(0, TILE_N)[:, None]
+            inp_offsets = pid_m * N * K + n_offsets * K + k_offsets
+            mask = (n_offsets < N) & (k_offsets < K)
+            inp = tl.load(input_ptr + inp_offsets, mask=mask, other=0).to(cdtype)
+            sum += inp
+        out = tl.sum(sum, axis=0, keep_dims=True)
+        out_offset = pid_m * K + k_offsets
+        output_ptrs = output_ptr + out_offset
+        tl.store(output_ptrs, out, mask=k_offsets < K)
+
+
+@libentry()
+@triton.heuristics(runtime.get_heuristic_config("softmax_inner"))
+@triton.jit
+def sum_dim_kernel_inner(
+    output_ptr,
+    input_ptr,
+    M,
+    N,
+    TILE_N: tl.constexpr,
+    ONE_TILE_PER_CTA: tl.constexpr,
+):
+    if tl.constexpr(input_ptr.dtype.element_ty == tl.float16) or tl.constexpr(
+        input_ptr.dtype.element_ty == tl.bfloat16
+    ):
+        cdtype = tl.float32
+    else:
+        cdtype = input_ptr.dtype.element_ty
+
+    pid_m = tle.program_id(0)
+    if ONE_TILE_PER_CTA:
+        n_offsets = tl.arange(0, TILE_N)
+        inp_offset = pid_m * N + n_offsets
+        input_ptrs = input_ptr + inp_offset
+        mask = n_offsets < N
+        inp = tl.load(input_ptrs, mask=mask, other=0)
+        out = tl.sum(inp, axis=0)
+        out_offset = pid_m
+        output_ptrs = output_ptr + out_offset
+        tl.store(output_ptrs, out)
+    else:
+        sum = tl.zeros(
+            [
+                TILE_N,
+            ],
+            dtype=cdtype,
+        )
+        for start_n in range(0, N, TILE_N):
+            n_offsets = start_n + tl.arange(0, TILE_N)
+            inp_offsets = pid_m * N + n_offsets
+            mask = n_offsets < N
+            inp = tl.load(input_ptr + inp_offsets, mask=mask, other=0).to(cdtype)
+            sum += inp
+        out = tl.sum(sum, axis=0)
+        out_offset = pid_m
+        output_ptrs = output_ptr + out_offset
+        tl.store(output_ptrs, out)
+
+
+@libentry()
+@libtuner(
+    configs=runtime.get_tuned_config("naive_reduction"),
+    key=["M", "N"],
+)
+@triton.jit
+def sum_dim_kernel(
+    inp,
+    out,
+    M,
+    N,
+    BLOCK_M: tl.constexpr,
+    BLOCK_N: tl.constexpr,
+):
+    if tl.constexpr(inp.dtype.element_ty == tl.float16) or tl.constexpr(
+        inp.dtype.element_ty == tl.bfloat16
+    ):
+        cdtype = tl.float32
+    else:
+        cdtype = inp.dtype.element_ty
+
+    # Map the program id to the row of inp it should compute.
+    pid = tle.program_id(0) * BLOCK_M + tl.arange(0, BLOCK_M)[:, None]
+    inp = inp + pid * N
+    out = out + pid
+    row_mask = pid < M
+
+    _sum = tl.zeros([BLOCK_M, BLOCK_N], dtype=cdtype)
+    for off in range(0, N, BLOCK_N):
+        cols = off + tl.arange(0, BLOCK_N)[None, :]
+        col_mask = cols < N
+        mask = row_mask and col_mask
+
+        a = tl.load(inp + cols, mask, other=0).to(cdtype)
+        _sum += a
+    sum = tl.sum(_sum, axis=1)[:, None]
+    tl.store(out, sum, row_mask)
+
+
+def sum_dim_comm(inp, dim=None, keepdim=False, *, dtype=None, out=None):
     if dtype is None:
         dtype = inp.dtype
         if dtype is torch.bool:
@@ -152,49 +251,125 @@ def sum_dim(inp, dim=None, keepdim=False, *, dtype=None):
 
     shape = list(inp.shape)
     dim = [d % inp.ndim for d in dim]
-    inp = dim_compress(inp, dim)
-    N = 1
-    for i in dim:
-        N *= shape[i]
-        shape[i] = 1
-    M = inp.numel() // N
 
-    out = torch.empty(shape, dtype=dtype, device=inp.device)
+    if len(dim) == 1:
+        dim = dim[0]
+        N = inp.shape[dim]
+        M = reduce(lambda x, y: x * y, shape[:dim], 1)
+        inp = inp.contiguous()
+        K = inp.numel() // M // N
+        shape[dim] = 1
+        if out is None:
+            out = torch.empty(shape, dtype=dtype, device=inp.device)
 
-    grid = lambda meta: (triton.cdiv(M, meta["BLOCK_M"]),)
-    with torch_device_fn.device(inp.device):
-        sum_kernel[grid](inp, out, M, N)
-    if not keepdim:
-        out = out.squeeze(dim=dim)
-    return out
+        with torch_device_fn.device(inp.device):
+            if K > 1:
+                grid = lambda meta: (M, triton.cdiv(K, meta["TILE_K"]), 1)
+                sum_dim_kernel_non_inner[grid](
+                    out,
+                    inp,
+                    M,
+                    N,
+                    K,
+                )
+            else:
+                grid = (M, 1, 1)
+                sum_dim_kernel_inner[grid](
+                    out,
+                    inp,
+                    M,
+                    N,
+                )
+        if not keepdim:
+            out = out.squeeze(dim=dim)
+        return out
+    else:
+        inp = dim_compress(inp, dim)
+        N = 1
+        for i in dim:
+            N *= shape[i]
+            shape[i] = 1
+        M = inp.numel() // N
+        if out is None:
+            out = torch.empty(shape, dtype=dtype, device=inp.device)
+
+        grid = lambda meta: (triton.cdiv(M, meta["BLOCK_M"]),)
+        with torch_device_fn.device(inp.device):
+            sum_dim_kernel[grid](inp, out, M, N)
+        if not keepdim:
+            out = out.squeeze(dim=dim)
+        return out
+
+
+def sum_dim(inp, dim=None, keepdim=False, *, dtype=None):
+    logger.debug("GEMS SUM_DIM")
+    return sum_dim_comm(inp, dim, keepdim, dtype=dtype)
 
 
 def sum_dim_out(inp, dim=None, keepdim=False, *, dtype=None, out):
     logger.debug("GEMS SUM_DIM_OUT")
-    if dtype is None:
-        dtype = inp.dtype
-        if dtype is torch.bool:
-            dtype = torch.int64
+    return sum_dim_comm(inp, dim, keepdim, dtype=dtype, out=out)
 
-    if dim == []:
-        if not keepdim:
-            return sum_out(inp, dtype=dtype, out=out)
-        else:
-            dim_num = inp.ndim
-            return torch.reshape(sum_out(inp, dtype=dtype, out=out), [1] * dim_num)
 
-    shape = list(inp.shape)
-    dim = [d % inp.ndim for d in dim]
-    inp = dim_compress(inp, dim)
-    N = 1
-    for i in dim:
-        N *= shape[i]
-        shape[i] = 1
-    M = inp.numel() // N
+# def sum_dim(inp, dim=None, keepdim=False, *, dtype=None):
+#     logger.debug("GEMS SUM DIM")
+#     if dtype is None:
+#         dtype = inp.dtype
+#         if dtype is torch.bool:
+#             dtype = torch.int64
 
-    grid = lambda meta: (triton.cdiv(M, meta["BLOCK_M"]),)
-    with torch_device_fn.device(inp.device):
-        sum_kernel[grid](inp, out, M, N)
-    if not keepdim:
-        out.squeeze_(dim=dim)
-    return out
+#     if dim == []:
+#         if not keepdim:
+#             return sum(inp, dtype=dtype)
+#         else:
+#             dim_num = inp.ndim
+#             return torch.reshape(sum(inp, dtype=dtype), [1] * dim_num)
+
+#     shape = list(inp.shape)
+#     dim = [d % inp.ndim for d in dim]
+#     inp = dim_compress(inp, dim)
+#     N = 1
+#     for i in dim:
+#         N *= shape[i]
+#         shape[i] = 1
+#     M = inp.numel() // N
+
+#     out = torch.empty(shape, dtype=dtype, device=inp.device)
+
+#     grid = lambda meta: (triton.cdiv(M, meta["BLOCK_M"]),)
+#     with torch_device_fn.device(inp.device):
+#         sum_kernel[grid](inp, out, M, N)
+#     if not keepdim:
+#         out = out.squeeze(dim=dim)
+#     return out
+
+
+# def sum_dim_out(inp, dim=None, keepdim=False, *, dtype=None, out):
+#     logger.debug("GEMS SUM_DIM_OUT")
+#     if dtype is None:
+#         dtype = inp.dtype
+#         if dtype is torch.bool:
+#             dtype = torch.int64
+
+#     if dim == []:
+#         if not keepdim:
+#             return sum_out(inp, dtype=dtype, out=out)
+#         else:
+#             dim_num = inp.ndim
+#             return torch.reshape(sum_out(inp, dtype=dtype, out=out), [1] * dim_num)
+
+#     shape = list(inp.shape)
+#     dim = [d % inp.ndim for d in dim]
+#     inp = dim_compress(inp, dim)
+#     N = 1
+#     for i in dim:
+#         N *= shape[i]
+#         shape[i] = 1
+#     M = inp.numel() // N
+
+#     grid = lambda meta: (triton.cdiv(M, meta["BLOCK_M"]),)
+#     with torch_device_fn.device(inp.device):
+#         sum_kernel[grid](inp, out, M, N)
+#     if not keepdim:
+#         out.squeeze_(dim=dim)
+#     return out