### PR Category
<!-- [ Operator | OP Test | Model Test | Benchmark | CI/CD | User Experience | Other] -->
Operator
### Type of Change
<!-- [ Bug Fix | New Feature | Performance Optimization | Refactor | Documentation Update | Other] -->
Performance Optimization
### Description
<!-- Briefly describe the changes and the purpose of the changes.-->
Opt sum dim for middle dim case

### Issue

<!--
List any related issues that this PR resolves, if applicable, for example:
- Resolves #123
- Associated with Feature #456
-->

### Progress

- [ ] Change is properly reviewed (1 reviewer required, 2 recommended).
- [ ] Change is responded to an issue.
- [ ] Change is fully covered by a UT.

### Performance
<!-- Please describe any performance tests you have added or the results of any benchmarks. -->

**Shapes in qwen3**
- before
```
benchmark/test_reduction_perf.py 
Operator: sum  Performance Test (dtype=torch.float16, mode=cuda,level=comprehensive)
Status       Torch Latency (ms)    Gems Latency (ms)         Gems Speedup          Torch GBPS            Gems GBPS           Size Detail
-----------------------------------------------------------------------------------------------------------------------------------------
SUCCESS               0.019328            0.015136               1.277             217.007             277.108          [torch.Size([1048576])]
SUCCESS               0.010656            0.007616               1.399               1.538               2.151          [torch.Size([64, 64]), 1]
SUCCESS               0.050912            0.043744               1.164            1318.135            1534.127          [torch.Size([4096, 4096]), 1]
SUCCESS               0.050752            0.224096               0.226            1322.290             299.465          [torch.Size([64, 512, 512]), 1]
SUCCESS               1.494208           13.030944               0.115            2874.411             329.598          [torch.Size([1024, 1024, 1024]), 1]
SUCCESS               0.010720            0.037856               0.283               6.113               1.731          [torch.Size([1, 8, 2048]), 1]
SUCCESS               0.011168            0.102112               0.109             281.673              30.807          [torch.Size([48, 8, 2048]), 1]
SUCCESS               0.443904           30.461409               0.015            2357.739              34.359          [torch.Size([15970, 8, 2048]), 1]
SUCCESS               0.045696            1.912704               0.024            1434.174              34.264          [torch.Size([1000, 8, 2048]), 1]
SUCCESS               0.043232            1.783488               0.024            1412.832              34.247          [torch.Size([932, 8, 2048]), 1]
SUCCESS               0.037536            1.530528               0.025            1395.015              34.213          [torch.Size([799, 8, 2048]), 1]
SUCCESS               0.034624            1.342304               0.026            1324.954              34.176          [torch.Size([700, 8, 2048]), 1]
SUCCESS               0.031872            1.193696               0.027            1278.972              34.149          [torch.Size([622, 8, 2048]), 1]
SUCCESS               0.040192            1.621792               0.025            1381.096              34.227          [torch.Size([847, 8, 2048]), 1]
SUCCESS               0.455520           31.200191               0.015            2357.178              34.415          [torch.Size([16384, 8, 2048]), 1]
SUCCESS               0.023488            0.728192               0.032            1054.692              34.019          [torch.Size([378, 8, 2048]), 1]
SUCCESS               0.020256            0.580064               0.035             970.616              33.894          [torch.Size([300, 8, 2048]), 1]
SUCCESS               0.016448            0.392384               0.042             800.872              33.571          [torch.Size([201, 8, 2048]), 1]
SUCCESS               0.012064            0.140384               0.086             369.401              31.745          [torch.Size([68, 8, 2048]), 1]


Operator: sum  Performance Test (dtype=torch.float32, mode=cuda,level=comprehensive)
Status       Torch Latency (ms)    Gems Latency (ms)         Gems Speedup          Torch GBPS            Gems GBPS           Size Detail
-----------------------------------------------------------------------------------------------------------------------------------------
SUCCESS               0.017536            0.014976               1.171             478.365             560.137          [torch.Size([1048576])]
SUCCESS               0.008928            0.007136               1.251               3.670               4.592          [torch.Size([64, 64]), 1]
SUCCESS               0.076064            0.072800               1.045            1764.537            1843.650          [torch.Size([4096, 4096]), 1]
SUCCESS               0.072448            0.264256               0.274            1852.608             507.908          [torch.Size([64, 512, 512]), 1]
SUCCESS               2.905344           28.224545               0.103            2956.598             304.343          [torch.Size([1024, 1024, 1024]), 1]
SUCCESS               0.009664            0.038048               0.254              13.563               3.445          [torch.Size([1, 8, 2048]), 1]
SUCCESS               0.012832            0.092960               0.138             490.294              67.679          [torch.Size([48, 8, 2048]), 1]
SUCCESS               0.857440           27.213600               0.032            2441.244              76.918          [torch.Size([15970, 8, 2048]), 1]
SUCCESS               0.075840            1.725344               0.044            1728.270              75.969          [torch.Size([1000, 8, 2048]), 1]
SUCCESS               0.072960            1.608928               0.045            1674.330              75.926          [torch.Size([932, 8, 2048]), 1]
SUCCESS               0.064320            1.381568               0.047            1628.211              75.803          [torch.Size([799, 8, 2048]), 1]
SUCCESS               0.058656            1.211904               0.048            1564.212              75.708          [torch.Size([700, 8, 2048]), 1]
SUCCESS               0.053408            1.079136               0.049            1526.490              75.548          [torch.Size([622, 8, 2048]), 1]
SUCCESS               0.067904            1.462688               0.046            1634.925              75.900          [torch.Size([847, 8, 2048]), 1]
SUCCESS               0.879264           27.874912               0.032            2442.365              77.040          [torch.Size([16384, 8, 2048]), 1]
SUCCESS               0.036096            0.661120               0.055            1372.596              74.941          [torch.Size([378, 8, 2048]), 1]
SUCCESS               0.030112            0.525440               0.057            1305.845              74.836          [torch.Size([300, 8, 2048]), 1]
SUCCESS               0.023424            0.353024               0.066            1124.721              74.628          [torch.Size([201, 8, 2048]), 1]
SUCCESS               0.014496            0.126816               0.114             614.852              70.282          [torch.Size([68, 8, 2048]), 1]


Operator: sum  Performance Test (dtype=torch.bfloat16, mode=cuda,level=comprehensive)
Status       Torch Latency (ms)    Gems Latency (ms)         Gems Speedup          Torch GBPS            Gems GBPS           Size Detail
-----------------------------------------------------------------------------------------------------------------------------------------
SUCCESS               0.016000            0.013024               1.229             262.144             322.044          [torch.Size([1048576])]
SUCCESS               0.008704            0.007136               1.220               1.882               2.296          [torch.Size([64, 64]), 1]
SUCCESS               0.050976            0.042912               1.188            1316.480            1563.872          [torch.Size([4096, 4096]), 1]
SUCCESS               0.050432            0.223776               0.225            1330.680             299.893          [torch.Size([64, 512, 512]), 1]
SUCCESS               1.492352           12.973152               0.115            2877.985             331.066          [torch.Size([1024, 1024, 1024]), 1]
SUCCESS               0.009664            0.039008               0.248               6.781               1.680          [torch.Size([1, 8, 2048]), 1]
SUCCESS               0.012224            0.100288               0.122             257.340              31.367          [torch.Size([48, 8, 2048]), 1]
SUCCESS               0.444544           29.853121               0.015            2354.345              35.059          [torch.Size([15970, 8, 2048]), 1]
SUCCESS               0.045280            1.875168               0.024            1447.350              34.949          [torch.Size([1000, 8, 2048]), 1]
SUCCESS               0.043264            1.748832               0.025            1411.787              34.926          [torch.Size([932, 8, 2048]), 1]
SUCCESS               0.038528            1.500448               0.026            1359.096              34.898          [torch.Size([799, 8, 2048]), 1]
SUCCESS               0.034624            1.316128               0.026            1324.954              34.856          [torch.Size([700, 8, 2048]), 1]
SUCCESS               0.031968            1.170432               0.027            1275.131              34.828          [torch.Size([622, 8, 2048]), 1]
SUCCESS               0.039584            1.589952               0.025            1402.309              34.912          [torch.Size([847, 8, 2048]), 1]
SUCCESS               0.455488           30.574944               0.015            2357.344              35.118          [torch.Size([16384, 8, 2048]), 1]
SUCCESS               0.022720            0.714016               0.032            1090.344              34.695          [torch.Size([378, 8, 2048]), 1]
SUCCESS               0.020288            0.568896               0.036             969.085              34.560          [torch.Size([300, 8, 2048]), 1]
SUCCESS               0.016544            0.384992               0.043             796.224              34.216          [torch.Size([201, 8, 2048]), 1]
SUCCESS               0.013152            0.137824               0.095             338.842              32.334          [torch.Size([68, 8, 2048]), 1]
```

- after 
```
benchmark/test_reduction_perf.py 
Operator: sum  Performance Test (dtype=torch.float16, mode=cuda,level=comprehensive)
Status       Torch Latency (ms)    Gems Latency (ms)         Gems Speedup          Torch GBPS            Gems GBPS           Size Detail
-----------------------------------------------------------------------------------------------------------------------------------------
SUCCESS               0.019392            0.015104               1.284             216.290             277.695          [torch.Size([1048576])]
SUCCESS               0.011200            0.007488               1.496               1.463               2.188          [torch.Size([64, 64]), 1]
SUCCESS               0.050144            0.043104               1.163            1338.323            1556.906          [torch.Size([4096, 4096]), 1]
SUCCESS               0.051520            0.062976               0.818            1302.579            1065.626          [torch.Size([64, 512, 512]), 1]
SUCCESS               1.494752            1.561664               0.957            2873.364            2750.251          [torch.Size([1024, 1024, 1024]), 1]
SUCCESS               0.010080            0.008960               1.125               6.502               7.314          [torch.Size([1, 8, 2048]), 1]
SUCCESS               0.011616            0.015776               0.736             270.810             199.400          [torch.Size([48, 8, 2048]), 1]
SUCCESS               0.444064            0.631488               0.703            2356.890            1657.371          [torch.Size([15970, 8, 2048]), 1]
SUCCESS               0.045760            0.064192               0.713            1432.168            1020.937          [torch.Size([1000, 8, 2048]), 1]
SUCCESS               0.043104            0.061856               0.697            1417.027             987.447          [torch.Size([932, 8, 2048]), 1]
SUCCESS               0.037504            0.053280               0.704            1396.205             982.794          [torch.Size([799, 8, 2048]), 1]
SUCCESS               0.034528            0.049184               0.702            1328.638             932.726          [torch.Size([700, 8, 2048]), 1]
SUCCESS               0.030912            0.044576               0.693            1318.691             914.469          [torch.Size([622, 8, 2048]), 1]
SUCCESS               0.039520            0.056416               0.701            1404.580             983.923          [torch.Size([847, 8, 2048]), 1]
SUCCESS               0.454944            0.647168               0.703            2360.163            1659.139          [torch.Size([16384, 8, 2048]), 1]
SUCCESS               0.022592            0.031456               0.718            1096.521             787.532          [torch.Size([378, 8, 2048]), 1]
SUCCESS               0.020224            0.027744               0.729             972.152             708.650          [torch.Size([300, 8, 2048]), 1]
SUCCESS               0.017216            0.020480               0.841             765.145             643.200          [torch.Size([201, 8, 2048]), 1]
SUCCESS               0.012032            0.014240               0.845             370.383             312.953          [torch.Size([68, 8, 2048]), 1]


Operator: sum  Performance Test (dtype=torch.float32, mode=cuda,level=comprehensive)
Status       Torch Latency (ms)    Gems Latency (ms)         Gems Speedup          Torch GBPS            Gems GBPS           Size Detail
-----------------------------------------------------------------------------------------------------------------------------------------
SUCCESS               0.017504            0.014848               1.179             479.240             564.966          [torch.Size([1048576])]
SUCCESS               0.009568            0.007488               1.278               3.425               4.376          [torch.Size([64, 64]), 1]
SUCCESS               0.076768            0.072352               1.061            1748.355            1855.066          [torch.Size([4096, 4096]), 1]
SUCCESS               0.072384            0.098624               0.734            1854.246            1360.903          [torch.Size([64, 512, 512]), 1]
SUCCESS               2.904928            3.015520               0.963            2957.022            2848.575          [torch.Size([1024, 1024, 1024]), 1]
SUCCESS               0.010528            0.008192               1.285              12.450              16.000          [torch.Size([1, 8, 2048]), 1]
SUCCESS               0.013728            0.011712               1.172             458.294             537.180          [torch.Size([48, 8, 2048]), 1]
SUCCESS               0.857568            0.855520               1.002            2440.879            2446.722          [torch.Size([15970, 8, 2048]), 1]
SUCCESS               0.075776            0.074208               1.021            1729.730            1766.279          [torch.Size([1000, 8, 2048]), 1]
SUCCESS               0.072160            0.070144               1.029            1692.892            1741.548          [torch.Size([932, 8, 2048]), 1]
SUCCESS               0.064288            0.063104               1.019            1629.021            1659.586          [torch.Size([799, 8, 2048]), 1]
SUCCESS               0.058560            0.056768               1.032            1566.776            1616.234          [torch.Size([700, 8, 2048]), 1]
SUCCESS               0.052704            0.051296               1.027            1546.880            1589.340          [torch.Size([622, 8, 2048]), 1]
SUCCESS               0.067872            0.065312               1.039            1635.696            1699.810          [torch.Size([847, 8, 2048]), 1]
SUCCESS               0.878304            0.877088               1.001            2445.035            2448.424          [torch.Size([16384, 8, 2048]), 1]
SUCCESS               0.035200            0.034944               1.007            1407.535            1417.846          [torch.Size([378, 8, 2048]), 1]
SUCCESS               0.030016            0.028384               1.057            1310.021            1385.344          [torch.Size([300, 8, 2048]), 1]
SUCCESS               0.024160            0.021152               1.142            1090.458            1245.531          [torch.Size([201, 8, 2048]), 1]
SUCCESS               0.014656            0.013280               1.104             608.140             671.152          [torch.Size([68, 8, 2048]), 1]


Operator: sum  Performance Test (dtype=torch.bfloat16, mode=cuda,level=comprehensive)
Status       Torch Latency (ms)    Gems Latency (ms)         Gems Speedup          Torch GBPS            Gems GBPS           Size Detail
-----------------------------------------------------------------------------------------------------------------------------------------
SUCCESS               0.016672            0.012864               1.296             251.578             326.050          [torch.Size([1048576])]
SUCCESS               0.008800            0.008224               1.070               1.862               1.992          [torch.Size([64, 64]), 1]
SUCCESS               0.050208            0.043680               1.149            1336.617            1536.375          [torch.Size([4096, 4096]), 1]
SUCCESS               0.051136            0.063808               0.801            1312.360            1051.731          [torch.Size([64, 512, 512]), 1]
SUCCESS               1.492736            1.560544               0.957            2877.245            2752.224          [torch.Size([1024, 1024, 1024]), 1]
SUCCESS               0.009888            0.007840               1.261               6.628               8.359          [torch.Size([1, 8, 2048]), 1]
SUCCESS               0.011808            0.015232               0.775             266.407             206.521          [torch.Size([48, 8, 2048]), 1]
SUCCESS               0.444704            0.632576               0.703            2353.498            1654.520          [torch.Size([15970, 8, 2048]), 1]
SUCCESS               0.045760            0.064000               0.715            1432.168            1024.000          [torch.Size([1000, 8, 2048]), 1]
SUCCESS               0.043232            0.060896               0.710            1412.832            1003.014          [torch.Size([932, 8, 2048]), 1]
SUCCESS               0.038432            0.053056               0.724            1362.491             986.943          [torch.Size([799, 8, 2048]), 1]
SUCCESS               0.034528            0.049152               0.702            1328.638             933.333          [torch.Size([700, 8, 2048]), 1]
SUCCESS               0.031840            0.045216               0.704            1280.257             901.526          [torch.Size([622, 8, 2048]), 1]
SUCCESS               0.040160            0.056224               0.714            1382.196             987.283          [torch.Size([847, 8, 2048]), 1]
SUCCESS               0.454816            0.647296               0.703            2360.827            1658.811          [torch.Size([16384, 8, 2048]), 1]
…